Face recognition/requirements.txt
dlib==19.24.4
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
pigpio==1.78